agent-backend/pyproject.toml
[project]
name = "agent-backend"
version = "0.1.0"
description = "Claude agent backend for the NovoEd toolkit course builder"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "claude-agent-sdk>=0.1.63",
    "python-dotenv>=1.0",
    "pydantic>=2.8",
]
